backend/requirements.txt
fastapi==0.111.0
uvicorn[standard]==0.30.1
python-dotenv==1.0.1
pydantic==2.7.1
langchain==0.2.11
langchain-openai==0.1.14
langchain-community==0.2.10
faiss-cpu>=1.9.0
numpy==1.26.4